requirements.txt
aiohttp
asyncpg
psycopg2
sqlalchemy-migrate
